Remove current subaccount mechanism
Previously the session instance held the current subaccount, which
defaulted to 0, persisted it on the server in the appearance data and
allowed it to be set by calling set_current_subaccount. The only place
where current subaccount was used was by create_transaction as the
default subaccount to use if it wasn't passed in the input json.

This change removes the current subaccount mechanism and leaves
determination of which subaccount to use to the client app. Having the
session hold the current subaccount doesn't add much value especially
in an environment where potentially multiple threads share a session.