Handle server appearance values of different types
The current implementation of applying defaults to the appearance
relies on the types of the server and default values being compatible.
For example for some wallets the value for "altimeout" is a string
instead of an int. This change simplifies the code and makes it agnostic
to the types of the values.