Transaction list caching and flexible paging
Change the interface to get_transactions so that the caller can specify
an arbitrary size set of transactions via 'first' and 'count' rather than
having fixed size pages from the server specified by page_id. This
allows the gdk to hide the details of the server page sizes from the
client and for clients to define their own page sizes.

Also implement a very simple cache for the transaction lists inside the
gdk.