Pick up transaction memos in cached list
Prior to commit 04d924f56fe2cef8ada802e8ff2c623ee46c4a09 every call to
get_transactions went to the server and so any transaction memos were
always up to date. There is currently no notification when a transaction
memo is updated so the cached tx list remains out of date.

This fix is a workaround which invalidates the local cache when a
transaction memo is set.

The real fix for this issue will be to trigger a notification to all
clients when transaction memos are updated but that has to wait until
such a mechanism is in place.