electrum: move sync from interface to lib [moveonly]
Due to borrow splitting issues[1] I had to move the sync function from
interface.rs

This is ok, since we're trying to phase this file out anyways. Now we
just do the sync from the ElectrumClient struct instead of WalletCtx.

We're also making a bunch of electrum_client calls in there as well,
so we no longer need to pass in client. We just get it from the
ElectrumSession directly.

This is mostly a moveonly, but I had to adjust a few lines where
client and wallet are accessed.

[1] https://doc.rust-lang.org/nomicon/borrow-splitting.html

Signed-off-by: William Casarin <jb55@jb55.com>