Replace GA_http_get with GA_http_request
This is a change to the external interface. It is assumed that
GA_http_get is only being called internally.

GA_http_request can perform POST as well as GET. urls are specified as a
set from which the gdk will choose the most appropriate, e.g. if the
session is running tor and an onion url is present it will get
preference over a non-onion url.