Get blinding nonces during create transaction
In liquid with a hardware wallet the blinding nonces for utxos must be
requested from the hardware wallet. Without this call create transaction
will fail because it cannot unblind the utxos it is trying to spend.

This issue is likely not noticed in apps because any call for example to
get_balance before create_transaction will fetch the blinding nonces as
a side effect.