Add change addresses to tx earlier
During create_transaction on liquid there will be two passes at
create_transaction. The hardware wallet needs to provide the blinding
nonces to unblind the utxos being spent and also a blinding key derived
for the change address.

Insert the change address into the tx json earlier so that it can be
seen by the hardware wallet during the resolver call, before create
transaction fails (due to not being able to calculate the fee because it
cannot yet unblind the utxos).