Key "satoshi" in transaction json on asset_tag
For liquid transactions with potentially multiple asset outputs the top
level "satoshi" field needs to be keyed on asset. Previously it
contained the value for btc outputs only and would be zero for a non-btc
output tx.

This is a breaking change for clients reading the top level "satoshi"
key from the transaction json and will require code changes to such
clients.