remove link argument if darwin
macosx jobs building ndk are failing due to this flag, it should not be necessary "to generate keys in gdk_rust" since key are not created in rust code, but through generate_mnemonic in cpp. One of the macosx building ndk will be run also when in branch, so we reduce the chance to discover this kind of errors only in master.