Use meson python module to resolve python
Previously the meson script created a virtualenv and then manually
constructed paths into the venv to find the python shared libs and
header file paths. Current versions of meson provide a python module
which provides these paths in a way which is more reliable, and
specifically fixes problems building the python wrappers in some
environments.