Fix inconsistent error handling enabling gauth
Fixes the following issues when enabling gauth as an additional
authentication factor:

- If an incorrect code is entered for the existing authentication
  factor there is no error from the gdk and is proceeds to request the
  code for gauth.
- If a correct code is entered for the existing authentication factor
  and then an incorrect code for gauth, the gdk reverts to asking for the
  existing.

The inconsistency arises because for gauth there is no call to request
the code and therefore nothing checks the code for the existing
authentication factor until the final call to enable_gauth where both
codes are checked and the gdk cannot distinguish which was wrong.

The backend api provides a proxy mechanism to improve this interface.