expose subaccount xpub fetching through session via an interface base class
This lets us instatiate an alternative implementation to derive via
different paths (e.g. BIP 44) without having to change any callers.

This is complicated a little by the fact that watch-only does not
(currently) create this class; in this case we continue to use the
static implementations for Green derivation.

Also add an as-yet unused method to remove subaccounts.