don't use the session cache for nlocktimes
locktimes are deleted whenevery the db is reloaded, and using the cache
requires serialising and re-parsing the data, and performing a DB
lookup. Instead of all that, just store them in the session. Also
simplify the map used, get rid of unneeded lambdas, and allow overriding
of the actual nlocktime data fetching.