Reimplement the tx cache.
This implementation avoids being limited by the server to 900 results.

The cache is populated on as as-needed basis. This should result in less
load for the backend since we no longer pass page numbers (other than the
degenerate case in local test envs where more than a page of txs can be
returned with the same creation timestamp). The server limits the DB query
by the timestamp we pass and thus doesn't need to read and throw away
result pages that wont be returned to the user.

We also make fewer calls to the server by determining the end of a
result set without requiring a round-trip returning no results.

Aside from better query performace, the cache also requests less data
from the server far less often; on new blocks and transactions only
mempool txs are reloaded instead of the whole cache, for example.