guard the tx list cache with the ga_session mutex
Also moves the munging of tx notification data outside of its handler,
so it can be done without the session mutex and doesn't have to be
re-done in the event the notification must be rescheduled (in an
upcoming change).

It is significantly easier to reason about the locking without separate
mutexes for the cache and its entries. This makes the deadlock between
notifications and the cache easier to understand and fix since it is now
based only on a single mutex (the ga_session mutex). The concurrency
advantage of finely grained cache threading is not useful in any case
given it takes 100s of milliseconds to fetch data from the server.