Implement a multi-call lock and use it to protect get_transactions calls
Some logical single level GDK calls (like get_transactions) are made up of
processing interleaved with multiple underlying WAMP calls. Since WAMP calls
require releasing the session mutex, its possible for a second thread to start
modifying the session in the middle of the logical call, causing inconsistent
results.

This is true for many calls implemented through auth handlers;
get_transactions is special because arriving notifications also change
the state used by the logical call, meaning even with an external caller
synchronising access to the session the session state can get corrupted.

To resolve this we need a lock that encompasses 'multi-call' calls and
works with the existing session mutex. We use a unique_ptr to a locker
instance to hold this lock, and either return it locked if no other
callers are competing, or unlocked if the caller elects not to wait.

Multiple external callers to get_transactions wait on the lock and so
are processed serially. Because the notification handlers are invoked by
the asio event loop, they cannot block without freezing the other
threads. So in this case if the lock cannot be taken, the handler is
re-posted to the event loop to be processed later when the lock is free.

As part of this change, the post login block notification is moved to
the very last action of the authenticate call.