rust: segregate db by network id only
Before this, any change to any of the the network fields (like the electrum
server url) would result in a separate db.

Now the db dir and the db encryption key are derived from an
unusual hardened path xpub (encoded with testnet version) and
the network id only.

This is a breaking change. Wallets created before this change will
loose the data in the db. However, before creating an empty db, the
wallet attempts recovery from the path used by Aqua wallet, which
currently is the only production wallet using GDK's electrum
networks. This should allow Aqua users not to loose their memos when
upgrading.

Suggested by @shesek.