Start simplifying the session networking implementation
To start, take advantage of the fact that the two transport alternatives
share a common base class. Hold a shared_ptr to the base transport instead
of a variant containing one or the other, and call through the base
where possible. This allows us to remove the template parameters from
most of the transport operations and avoid having to expose this code
through the session header.

As part of this change, remove the templated chain-of-deferred-futures from
the connection codepath; since we wait on each future synchronously anyway
there is no reason to run each in a different thread.