Use weak_ptr to prevent callbacks firing on dead/other sessions
As the ping/pong callbacks can happen on any thread, fetching the
session impl can race causing callbacks on a dead session or in the
worst case a different session than the one intended. Use a weak_ptr
bound to the created session to ensure we only process the callback
if (1) The session is still live and (2) its the original session we
set the callback on.

This prevents the handlers crashing, but is not a final fix.