rust: allow caller to recognize invalid pin errors
After 3 wrong pin attempts, pinserver deletes the encryption key.
Caller should keep a counter of the the remaining attempts, to
display a proper warning to the user.
This counter should be increased after a wrong attempt, but should
not be increased if other errors, such as connection issues, are
returned.