Use set_data() consistently to set auth_handler two factor data
We currently overload the auth_handler state machine to handle both HW
and user authorisation requests; the presence of HW device information
is used to indicate to the caller whether a device or human interaction
is expected.

Using set_data here ensures that the HW device info is set consistently
only when required for the current action (after the following patch, which
fixes cases that change the action without updating the HW device info).