Use an internal software implementation of the HWW interface (1)
By implementing the hardware wallet interface internally, we simplify
call handlers to a single path of execution. This is a requirement for
caching data correctly, since we need the data in its final format in
order to cache it (e.g. caching unblinded data is not a performance
win), and we need to encache and re-fetch that data at the same point
of execution regardless of whether keys are external or internal.

This initial change implements just enough of the HW interface for login
to succeed using the internal software interface. This requires some
restructuring including:

- Deriving the software signer from the hardware signer class
- Moving the responsibility of holding the mnemonic into the signer
- Implementing an auth_handler wrapper to automatically advance the
  state machine for the software interface
- Removing the SW/HW overrides of register_user and code paths in
  register/login auth_handlers
- Removing the login_with_pin method and replacing it with
  pin->mnemmonic conversion followed by standard login