Simplify create_transaction auth handler
Avoid unneeded JSON copies. Move directly to state done if no blinding
is required, in particular avoid repeated unblinding of the same
addresses.

Also remove code to encache blinding nonces - this can never be executed
since the handler never resolves a call to fetch them. Presumably this
code is a remnant of the initial Liquid support.