rust: elements: fix p2wpkh, p2pkh address generation
Before this change, we were always using the p2shwpkh script to
derive the blinding public key from the master blinding key. As
a result the address generated had an unexpected blinding public
key, which later caused an unblinding failure and made the fund
unspendable.

Funds are not definitively lost as the blinding key was still
deterministic, but it's not trivial to recover them. Anyway, this
never went live on a production app.