Remove the has_xxx cache functions for Liquid utxos and blinding nonces
They can initially be removed by just using the fetch functions; In due
course the inserts can avoid checking before inserting with the
appropriate ON CONFLICT clauses added.

It remains somewhat confusing that we cache both the blinding nonces and
the unblinded outputs however. The nonces are sufficient to unlind, and
if we cache both transactions and the utxos unblinded there seems little
point in also caching the outputs separately. Also, having the same
information cached in 2 different ways is just asking for them to get
out of sync.