rust: request exchange rate at most 1 time per minute
In case a price source is not responsive, or not reachable through
a proxy, attempting to get an exchange rate every few second can be
a waste of time and can degrade the UX with unnecessary delays.

With this change, we do at most a request per minute. This is suboptimal
in some cases, as we might obtain the rate a bit later. But if the price
provider is down for maintainance or is completely unreachable e.g.
through tor, this prevents the apps to slow down significantly.