Don't mark the users client blob outdated in watch only mode
If a full and watch only session are logged in at the same time, and the
full session updates its client blob, the watch only session will be
notified. Prior to this change, that would cause the watch only login to
mark its blob outdated and subsequently attempt to load the blob, which will
fail.