Use the correct tx/vout pair when unblinding UTXOs
It is ~20 times faster to read a cached unblinded UTXO than to unblind
it using a cached nonce (which is itself orders of magnitude faster
than asking the hardware to produce the nonce if it is not cached).

Prior to this change, full unblinded data was not cached for transaction
inputs and outputs; only the nonces were cached. This means the original
design using needs_unblind_call to fetch and cache the blinding data up
front from the users transaction list was significantly less efficient than
it should have been, since it only ever cached nonces.

It is also possible there is/was a code path which leads to full
unblinding data being cached against the wrong txhash/vout pair. That
should be fixed by this change if so.