Move notification handler implementation to session_impl
For session_impl to invalidate cached data, it needs to be able to
respond to notifications. Change multisig notification emission to
match single sig (provide the entire notification). Remove auto-resetting
of the notification handler to null since this path can't be reached.
Reduce copying of notification data.

Along with this change, move the lock type and mutex to session_impl to
be shared by both session types. Since the notification handler can only
be set prior to connect, remove locking from setting and calling through
it.