Remove "confidential_utxos_only" from tx creation
UTXOs can now be filtered by:
- num_confs
- user status
- expiry
- confidentiality
- dust limit

Of these, only confidentiality is currently available to override in
create_transaction. Instead of gdk exposing each via a different name,
the caller should fetch the UTXOs filtered as they wish and pass them in.
This removes any ambiguity if the filters are set
in the create_transaction json and UTXOs are already present (i.e does
the caller want the passed in UTXOs re-filtered in this case/should this
be an error/what happens during interactive editing etc).

As confidential_utxos_only is not yet used by any wallets, and coin control
requires the UTXOs to be passed in anyway, remove this option now so
that filter handling and coin control in the wallets can use the same
code paths.

Requires rust changes to remove the element from its create_transaction
code. Leaving that for now as there are outstanding test merges with the
potential to conflict.