Fetch fee estimates on demand instead of updating them via subscription
Fees are only required by clients when creating transactions, so have
the caller poll instead of updating the data redundantly from the server
every two minutes. The switch to polling for estimates is enabled by a
server upgrade; until this is detected by gdk the client will continue
to update via subscribing.

Polled data is cached and refreshed if requested after more than two
minutes have elapsed.

As a result of this change, fees are no longer notified to callers.