create_transaction: check change output blinding after recreating the tx
Currently, the create tx code always creates change addresses for each
asset in the tx, even if there is no change. This is sub-optimal since
each call requires a server round-trip, the output is blinded even if
unused, and if not used the address is thrown away.

Update the call handler to loop while there are unblinded change
outputs, which allows the create tx code to add change outputs on demand
(in a later change) instead of up-front.