rust: tests: sleep more, but less often
Wait 1s at most 60 times, instead of 0.5s at most 120 times.
This is more clear and has a minimal impact on test runs (at most
5 secs more on a dedicated machine for all integration tests; it's
reasonable to expect that this increase will be even more negligible
on a busy CI box).

Also remove reference to "1 minute" as we might tweak the delay
time in the future.