Automatically advance auth handlers where possible
Now that all auth handlers have their logic in the call() method, allow
auto_auth_handler to fast-forward each action's state machine until
caller interaction is needed.

This greatly reduces the amount of caller interation when resolving,
and reduces the amount of data that must be passed back and forth. In
many cases auth handlers can proceed directly to the done state.

This also means that auto_auth_handler will be able to handle
reconnecting/re-logging in the session automatically in the future.