Use the sessions signer in login_call when re-logging in
This makes the login invariants identical for both session types, and
enables using the session's signer's cached xpubs from the last login.

As a result we can remove passing back the relogin status in
post_login_data, and allow the normal flow to proceed when re-logging
in, so we e.g. upload confidential addresses if needed.