ga_session: inline disconnect_transport into its caller
This change makes it clear that disconnection does not reset m_session
or m_transport to null as one would expect.

Given that other access to these members is not always checked and
additionally is not mutexed, this seems likely to be a source of rare
crashes to be addressed in a future networking update.