Reimplement transaction fetching to use a persistent cache (2)
Use the new txs.get_list_v3 backend call to fetch transactions. Sync forward
and then return the requested data, restarting the sync if the cache is
invalidated while syncing. Note that invalidation is not part of this
commit.

Transactions are synced and cleaned up/unblinded before being cached. Tx
munging code is updated due to server data changes and to reduce copies.
Memos and spv status are grafted into the result data after loading from
the cache is complete.

In this initial implementation, cache invalidation will be handled by
directly clearing the cache, for simplicity.

As part of this change, we remove instant/has_payment_request from the
returned tx json.