Add get_max_reorg_blocks to network_parameters
Returns the maximum number of blocks we can expect in a re-org for the
network type. Since the current backend does not provide sufficient
information to be more precise, we err on the side of caution when a
re-org or missed block notification is detected.

Amortised over the lifetime of a wallet, these values have little effect
on performance. All transactions ultimately become old enough to avoid
re-org culling in due course.