Return wallet_hash_id when registering, add GA_get_wallet_identifier
Callers can now use the identifier prior to login, when registering and
when logging in to associate local data with the network + wallet,
without exposing the actual xpub.

This involves moving a fair amount of code around to expose the
functionality in multiple places, and allows sharing more code between
cpp and rust sessions.