Update wally to 0.8.4
Also updates the wally build code to handle libsecp as a submodule.

meson doesn't handle this case, so we hack it by cloning the correct
libsecp commit into the source tree before we then build wally.

To minimise the possibility of error with this approach, we store the
secp commit in the wrap file along with the wally version. Also update
the build script to handle non-release versions of wally which was lost
with the update to 0.8.2.