refresh_assets: add policy asset consistently
Do not add the policy asset in get_inbuilt_data for 2 reasons
(a) it was added at the wrong json level, instead of inside the
"body" field (b) it is added later in `ga_session::refresh_assets`
after the patch is computed and/or applied.

Also set the policy asset consistently for Liquid Electrum sessions.
Note that this is a breaking change.
The policy asset entry before was:
{"asset_id": "...", "name": "Liquid Bitcoin", "ticker": "L-BTC"}
now is:
{"asset_id": "...", "name": "btc"}