rust: implement get_raw_transaction_details
For the time being, transaction are obtained from the local store.
This means that (a) such transactions are stripped of the witness
and (b) if the tx does not belong to wallet an error is returned.

This is not exactly matching what "multisig" is doing, but it's
getting us closer to that.