Use AtomicBool to signal thread termination so that it could be checked before sending notification
Before this commit to signal thread termination a message was used, and threads terminated after
terminating the loop cycle, during this last loop they may emit notification on an handler deleted
from the cpp side causing a crash. With this commit an atomic bool is used and is checked both
to break the loop and right before emitting a notification.