net: partially re-work reconnection handling
- Use a member to hold any reconnection thread in progress.
- Make reconnect_hint wait for any reconnection thread to finish when
  disabling reconnect.
- Use reconnect_hint(false) in the ga_session dtor to ensure reconnections
  are stopped before we are deleted.
- Move reconnect_impl into the reconnect thread for now.
- Always unsubscribe when disconnecting.
- Update logging

reconnect() can be called from the users thread(s), or the ping/hearbeat
timers inside the asio run loop. If called from a users thread, submit
the reconnect to the run loop and wait for it to to complete; this ensures
that the reconnection thread is always initiated and acceessed from the
same context.