net: re-design wamp network logic
The design here uses the asio context running with a single thread to
process all aync operations in the background in a serial fashion. This
is more than enough concurrency for a single session object.

Reconnection (both internal and external) and ping timeout is handled via
a second, private thread which monitors the connection state and
responds to caller-initiated and internal requests to update the state.

We use a condition variable to make the reconnection thread aware of
desired changes and error conditions. This reduces the state management
variables/states considerably.

Because we only have a single thread updating the state, this prevents
repeated reconnections being in-flight as seen with the previous design.
Similarly, a single asio thread ensures that notifications are processed
in order (unless the session is locked when the notification is
processed, this will be fixed in a separate PR).

The implementation is complicated slightly by bugs in
autobahn/websocketpp which require the reconnction thread to post some actions
to the io context for thread safety, and to poll for state changes which not
correctly signalled by the autobahn returned future objects.