net: add initial connection ping logic
Pings appear to be useful only in the case where we are using a proxy
(including our internal tor proxy). In these cases, the tcp connection
to the proxy can remain open, while the proxy connection to the server
has dropped.

Without a proxy, the socket will be dropped when the connection is, and
is_connected() will return false before any missed pong is detected.