net: increase the ping delay to once every 60 seconds
The pong timeout on the connection is set to 20 seconds. With the ping
timer at 20 seconds also (and especially in the previous impl with
multiple asio context threads) it was not uncommon to have pings in flight
when pongs were already missing.

Since we avoid pinging when the connection is active (e.g. when making
wamp calls), we can lower the ping frequency without affecting the
response time, since the wamp call timeout is ~20s already (and when the
connection is idle there is no point pinging so frequently).