rust: fix account_nums bug
In 260364b73afe0edfc69c4b7a5f62432450daf944 we started reading
the account nums from the store rather than the cache.
However the store is not populated by previous versions of gdk
(e.g. 0.0.48), so when upgrading GDK from an old version it will
miss some accounts.

This fix looks both at the store and the cache to return account
nums fixing the above issue.

Note that this fix will not work if the cache needs to be
reconstructed.