auth handlers: use a chained handler for confidential address upload
This is the simplest possible chained handler case, since (1) it always
happens at the end of the parent auth handler actions, and (2) the caller
does not need access to its result (other than success or failure).

The logic for the new handler is written in sequentual order to make it
easier to follow. It has also been made re-startable since intrusive
chained auth handlers (re-login) will require this.