python: Fix session/auth_handler lifetime issues
Has fixes for the session lifetime, reference counts, and thread safety.

Makes attempting to destroy a session multiple times, and accessing it
after it has been destroyed, safe (in that it non longer crashes),

For explicit lifetime control, the caller whould use a with-block or
call destroy() themselves. Otherwise, the garbage collector will
dispose of the session at some point; if not then it will be deleted at
program exit.