spv: wait longer when waiting for the current block to arrive
When initially catching up with headers, sleep for a short period (50ms)
to allow progress by other threads and to allow cancelling the sync in a
reasonable time.

When we have not synced to the current block after the session has seen it,
increase the delay to 1s to allow the upstream electrum server time to
receive and process the block before looping.