ga_rust: go back to the cpp impl of get_subaccount_root_path
This method is currently exposed from rust as a method on the session,
but is called before the session is created for fetching subaccount
xpubs.

Therefore it can either be left as is in the cpp or the rust impl needs
to be made stateless/global rather than a session method.