rust: panic if download_headers called in liquid networks
Instead of `get_bip32_network` use NetworkId and panic if `download_headers` is
called in Liquid networks

tests: avoid to download headers chain in liquid

wait on the handle before finishing the test so that panic! in the thread
make the test fail

also remove a println! in another test