rename "public_key" to "blinding_key" when appropriate
"public_key" is used for sweep and single sig for the public key
corresponding to an output.

"blinding_key" is used by blind_addres for the blinding public key
corresponding to an output.

However "public_key" was used in some places for the blinding
public key, this commit rename the key in such places.

This is a breaking change.