rust: wallet_hash_id: overwite fingerprint, child number and depth
This makes the wallet_hash_id consistent with multisig even in the
case where the above fields are set (which is odd, since that
should be a master key).

There is evidence this is needed for some hardware wallets.

This should not affect software wallets and their existing cache,
because their master xpub already has the value set to 0.