rust: use LTO in Darwin for debug builds
For obscure reasons link time optimizations of the rust library are required
for debug builds in iphone and iphonesim environment.
Instead of placing the setting in the Cargo.toml, ruining the time of rust devs
CARGO_PROFILE_DEV_LTO env var overwrites that value only in CI and only in
Darwin build.