rust: create_transaction: make is_change equivalent to is_internal
Before this commit is_change was the same as is_relevant. However this
causes issues in case the user does a redeposit/send to self.

It's worth noting that sometimes using is_internal for is_change is
incorrect, for instance if the user sends funds to one of its addresses
belonging to the internal change.