rust: store recently spent utxos in memory
If a caller sends/broadcasts a transaction, and *before* such
transction is synced (before the notification is received) it
attempts to send another transcation, it might use the same
utxos used by the first transaction. This results in a
"double spend" error.

To prevent this, store the recently spend utxos in memory.
Then at the next sync, remove all of them.