rust: get_unspent_outputs: fix bug returning utxos from replaced txs
In 0a55d9944e0f02df2948cf692836eadb703d6ac0 a regression was
introduced, which caused transactions that have been replaced or
dropped out of mempool be included in the utxos. This might cause
the balance to be incorrect or sent transactions to be rejected.

This commit fixes the above issue.