[build-system] separating build of external dependencies
This commit aims at removing the dependencies build step from the build process
of the gdk library, and thus making the iterative build of gdk faster both
locally and in CI.
The newly introduced script, ``tools/builddeps.sh``, downloads and builds all
the external dependencies eventually installing them in a single user-defined
folder, which can be then passed to ``tools/build.sh`` via the
``--external-deps-dir`` optional argument.
In the optimal path where ``--external-deps-dir`` is set correctly,
``tools/build.sh`` becomes leaner and quicker in building gdk as it no longer
needs to build all the libraries gdk depends upon, however, the script is still
capable of handling the case where the option is not passed, or the folder does
not contain the expected dependencies (a boost library is sampled to assess
well-formed presence of dependencies) and call itself ``builddeps.sh`` script.

As a result of this change, all subprojects ``.wrap`` and ``meson.build`` files
are removed as they no longer serve any purpose.

The Dockerfile for debian CI image is changing accordingly and pre-building all
the dependencies for all the supported platforms, gitlab-ci.yml is also
changing to leverage those pre-built libraries.