create_transaction: only remove change for multi-asset sends
The check introduced in 0ece9e297f085bdfd7e5344e09c32430c208c8da was
incorrectly applying to all transactions.

A multi-asset send can be explicitly multiple assets, or a single
non-L-BTC asset with L-BTC required for fees. Esentially, any tx that
isn't a simple L-BTC send.

Also, "change_subaccount" is set by the caller to direct any change
to a specific subaccount, and so should not be deleted.