rust: don't strip witness, breaks cache
To be coherent with the interface we are not going to strip the witness from
the transactions before saving in the cache, to allow retrieve the witness
when transaction details are asked.

Since all the transactions must be re-downloaded we took the occasion to simply
break the cache compatibility (triggering a cache rebuild) and remove
compatibility-preserving optional from the cache.

Along the lines of the test checking for the compatibility of the store adds
the same test for the cache. Breaking the cache is not critical but better to
be warned by a test if someone is doing it inadvertently.