create_transaction: do not require or return top-level "subaccount"/"subaccount_type"
Cross-subaccount transactions require that we infer the affected subaccounts,
so we don't need the user to specify them. Gather the subaccounts from
the inputs passed to create_transaction instead, which matches the
existing behaviour for non-cross-subaccount txs.

Sweep and bump currently still require the top-level subaccount, however
these are not available on Liquid where cross-subaccount spends are
expected via swaps. Full removal of these elements in these cases requires
refactoring that will be done after the single and multisig creation
code is merged.