create_transaction: add support for blinding with scalars
Callers might have already blinded some outputs and might not be
willing to share the corresponding amount blinder (value blinding
factor or vbf). They might also not want to share the vbf for some
inputs as well.

For such inputs and outputs, their contribution to the last vbf
does not come from its value, abf and vbf, but comes from the
scalar offsets.

Additionally this commit solves an issue where partial blinding
generated the last vbf, which wasn't actually the last one, so that
it balanced the previous ones. Now it is generated randomly.