rust: (gdk_rust) separate `GDKRUST_call_session` into safe and unsafe parts
This splits the implementation of the `GDKRUST_call_session` into an
unsafe exposed wrapper which is responsible of handling raw pointers, and
a safe Rust function where the logic is implemented.

An added benefit of this is being able to use the `?` operator to bubble
up errors instead of having to `match` and `return GA_ERROR` at every
step.