[build-system] adding cargo-audit to docker CI
cargo-audit is a nice tool to check project's dependencies against
RUSTSEC DB of known security issues.
This commit adds cargo audit to the Debian docker CI image and adds a
gitlab-ci extra stage named ``audit`` with an extra job that launches
cargo-audit.
The job and the stage are conditionale to the MR containing changes to
gdk_rust.
For C++ only MRs this commit will not affect their CI flow in any way.