build: prevent dependency rebuilds, simplify logic
Simplify the dependency logic so that we always have one directory with
the built dependencies, and a link to it (even in the case where no
dependency dir is passed to the build script). If we need to build the
dependencies, do it into the build directory and not through the link.
This prevents the dependencies from being rebuilt every time because the
link has been removed, and allows to recreate the link the same way for
every build.

Also share the various repeats of this logic in a single place.