[bug-fix #754] adding exported symbol for iOS
It seems the cause is due to a change in the rust compiler 1.62 and
above, due to adoption of LLVM 14 that "perfectly" aligns with some weird
stuff we do in our scripts to eventually mess things up.
In bundling libgdk_rust.a for release mode, the new compiler no longer resolves
internal symbols, but rather leaves ``_rust_eh_personality`` function as
undefined, even if that same function is defined within the library
itself.
Running ``nm`` onto libgdk_rust.a artifact, we get
```
[...]
T _rust_eh_personality
U _rust_eh_personality
[...]
```
In rust 1.59 and below this does not happen as all the ``U`` symbols are
resolved.

Then it comes our black magic trick: in our tools/buildgdk_rust.a
script, we happen to remove all symbols except for a few defined in the
apple-exported-symbols file.
The hand is faster than the eye, yet if you follow the traces you can notice
that the trick applies only to iOS systems.
The final linker is then left with a bunch of ``U _rust_eh_personality``
with no ``T _rust_eh_personality`` as the ``T _rust_eh_personality`` is
stripped away in the previous stage.

This commit solves the issue by simply adding ``_rust_eh_personality``
to the list of exported symbols for apple.

A further investigation is needed to eventually remove completely the
intermediate stripping of symbols.
IMHO stripping should, in principle, happen only at the very last stage,
against libgreenaddress.[so|a]