rust: ask read lock once in create_tx
This avoids a deadlock on the Store in the following situation:

One thread is in the middle of create_tx, after having asked and obtained
 one read lock, and before asking another read lock.

Another thread in sync ask for the write lock.

This results in a deadlock. I am not sure why it happens right now but I presume
the updates in rust 1.62 on the Mutex it's related:
https://blog.rust-lang.org/2022/06/30/Rust-1.62.0.html#thinner-faster-mutexes-on-linux

As a consequence of making one lock asking in `create_tx` also `get_unspent_outputs`
has been updated accordingly.

This doesn't remove any possible occurence, and a more involved change preventing
this kind of deadlock is necessary, but in the meantime this looks improving
test reliability a lot.