electrum: login_user: prevent AccountGapsDisallowed
Subaccounts created at login come from the cache/store so in theory
they can't have gaps. But if for any reason we see a gap here, we
should not error and prevent login since it might lock a user out of
its wallet.

GA_create_subaccount will still error if there is a gap.