[build-system][ios]builddeps going solo
iphone and macOS builds were the only builds that are still relying on
build.sh internally calling builddeps.sh.
With the aim of battle-testing builddeps.sh, as part of the project of
reducing the span of build.sh, this commit changes the CI scripts and
splits the dependency building part from the gdk one.