rust: create_account: allow_gaps
We want to allow users to create watch-only wallets for any subset
of their accounts. For instance a user might have 2 P2PKH accounts
and it wants to create a watch-only wallet for the 2nd one. Such a
wallet would have a gap (account 0 P2PKH is missing), thus we must
allow to relax the gaps checks.