addresses: only check confidential addresses that are the correct length
Addresses are often checked to see if they are confidential. While the code
to do this was changed to prevent log spam when the address isn't
confidential, it still throws an exception. This makes debugging some
test cases much more painful. Instead of checking unconditionally, only
check when the base58 address can actually be a confidential address.