[build-system] dev-mode for faster python-wheel dev process
gdk development-and-testing loop typically involves the e2e tests
running in python environment using the gdk python wheel.

Having developers re-building the python wheel, uninstalling and
re-installing the wheel every time they make a change to the c++ or rust
code introduces an additional burden on their development-testing loop.

This commit introduces a devmode option in the build-system that allows
developers to shortcut some of the python wheel building process.

With devmode activated, developers no longer need to re-build or
re-install the wheel as the wheel itself is pointing to
libgreenaddress.so dynamic library, an artifact that is already rebuilt
every time a change to C++ or Rust code is detected.
Once the libgreenaddress.so is rebuilt, devs will have to restart their
python environment, with no need to re-install anything, but only allow
the python interpreter to unload the old libgreenaddress.so and re-load
the new one.