build-system: preparing env for giulio_cibuildwheel
This set of changes paves the way for cibuildwheel CI job.
``cibuildwheel`` python package is a convenient tool to build
pypi.org-compatible python wheels that support different python
versions and different glibc versions.
This commit also improves in general the overall machinery that builds
the wheel both in local devs environment and in the CI.