build: update ubuntu build, allow building with provided sources
Use the latest ubunutu image to minimise the number of updated packages.

Use the requirements.txt file from its in-source position to avoid
special handling and make the deps script runnable locally.

If the dependency source tarballs are available in the build directory,
they will be copied into the base image and then used from there
provided they pass the sha256sum check. This makes the build *much*
faster. Although the resulting image is larger, it can be cached and the
resulting dependency dependency build images are then smaller.