build-system: extract rust from dependencies
This commit refactors the way rust and its tools are included in the
final docker image.
Rather than manually installing rust via rustup during the early stages
of the CI docker image building process, we take rust straight from an
official docker image for debian and we use that image as base to build
the tools we use in our current CI pipeline.
This results in a faster, more parallelizable, more re-usable building process.